Updated the java docs for Manga.java

+++ src/MyAnimeList/Manga.java
@@ -22,10 +22,18 @@ public class Manga extends MyAnimeListMedia {
         super(id, "manga", MANGA_PROPERTY_TYPES);
     }
    
+    /**
+     * Gets the number of chapters in the manga
+     * @return Number of chapters
+     */
     public int getChapters() {
         return getIntegerValue("chapters");
     }
     
+    /**
+     * Gets list of user-submitted manga recommendations based on this manga
+     * @return List of manga recommendations, sorted in descending order of recommendation count
+     */
     public TreeMap<Integer, Manga> getRecommendations() {
         if (recommendations == null) {
             setRecommendations();
@@ -34,18 +42,34 @@ public class Manga extends MyAnimeListMedia {
         return recommendations;
     }
     
-    public double getScoredBy() {
-        return getDoubleValue("scored_by");
+    /**
+     * Gets the number of users who scored this manga
+     * @return Number of scorers
+     */
+    public int getScoredBy() {
+        return getIntegerValue("scored_by");
     }
     
+    /**
+     * Gets the number of volumes for this manga
+     * @return Number of volumes
+     */
     public int getVolumes() {
         return getIntegerValue("volumes");
     }
     
+    /**
+     * Gets the publishing status of this manga
+     * @return True if still being published, False if not
+     */
     public boolean isPublishing() {
         return getBooleanValue("publishing");
     }
     
+    /**
+     * Updates the recommendations based on this manga
+     * @return True if recommendations is successfully updated, False if not
+     */
     private boolean setRecommendations() {
         TreeMap<Integer, Manga> recommendations = new TreeMap<>();
